Statistics.stack: Move h/vstack to statistics module.
